Content: Show block references even if content item has no fields (#4059)
Moved the block references logic to the Content component so even if a
content item has no fields, we can still show the block references. This
addresses the comment
[here](https://github.com/zesty-io/manager-ui/issues/4005#issuecomment-4392795287)
Resolves #4005

### Preview


[Screencast_20260507_111106.webm](https://github.com/user-attachments/assets/9a243177-6754-40f9-a84a-8e308bba3f6f)
