+++ src/apps/content-editor/src/app/components/Editor/Editor.js
@@ -441,7 +441,6 @@ export default memo(function Editor({
           </div>
         );
       })}
-      {model.type !== "block" && <UsedBlocks />}
     </div>
   );
 });
