+++ src/apps/content-editor/src/app/views/ItemEdit/Content/Content.js
@@ -13,14 +13,34 @@ import { theme } from "@zesty-io/material";
 import { StartRounded, DesktopMacRounded } from "@mui/icons-material";
 import { Actions } from "./Actions";
 import { useLocalStorage } from "react-use";
-import { useContext, useEffect, useMemo } from "react";
+import { useContext, useEffect, useMemo, useState } from "react";
 import { DuoModeContext } from "../../../../../../../shell/contexts/duoModeContext";
 import { FieldError } from "../../../components/Editor/FieldError";
 import { BlockTabs } from "../components/BlockTabs";
-import { useGetContentModelFieldsQuery } from "../../../../../../../shell/services/instance";
+import {
+  useGetContentModelFieldsQuery,
+  useGetContentModelsQuery,
+  useGetWebViewsQuery,
+} from "../../../../../../../shell/services/instance";
+import {
+  fetchItems,
+  searchItems,
+} from "../../../../../../../shell/store/content";
+import { fetchVersions } from "../../../../../../../shell/store/contentVersions";
+import { useDispatch, useSelector, useStore } from "react-redux";
 import { NoFields } from "../../../components/NoFields";
+import { UsedBlocks } from "../../../components/UsedBlocks";
+import { extractBlockReferences } from "../../../components/UsedBlocks/extractBlockReferences";
 
 export default function Content(props) {
+  const dispatch = useDispatch();
+  const store = useStore();
+  const { data: views } = useGetWebViewsQuery({ status: "dev" });
+  const { data: models } = useGetContentModelsQuery();
+  const [blockReferences, setBlockReferences] = useState([]);
+  const [isBuildingReferences, setIsBuildingReferences] = useState(false);
+  const contentItems = useSelector((state) => state.content);
+  const contentVersions = useSelector((state) => state.contentVersions);
   const [showSidebar, setShowSidebar] = useLocalStorage(
     "zesty:content:sidebarOpen",
     false
@@ -70,7 +90,195 @@ export default function Content(props) {
       !Object.keys(props.item?.web).length ||
       !Object.keys(props.item?.meta).length);
 
-  if (!hasFields && !isFetchingFields) {
+  const blockModels = useMemo(() => {
+    if (!models) return [];
+    return models.filter((model) => model.type === "block");
+  }, [models]);
+
+  const templateFiles = useMemo(() => {
+    if (!views || !props.modelZUID) return null;
+    return views?.reduce(
+      (acc, view) => {
+        if (view.contentModelZUID === props.modelZUID) {
+          acc.code = view;
+        }
+        if (view.fileName === `/z/pvl/${props.itemZUID}.zhtml`) {
+          acc.freestyle = view;
+        }
+        return acc;
+      },
+      { code: null, freestyle: null }
+    );
+  }, [views, props.modelZUID, props.itemZUID]);
+
+  useEffect(() => {
+    const run = async () => {
+      if (!templateFiles) return;
+
+      setIsBuildingReferences(true);
+
+      const allBlockReferences = [
+        ...(templateFiles.code?.code
+          ? extractBlockReferences(templateFiles.code.code)
+          : []),
+        ...(templateFiles.freestyle?.code
+          ? extractBlockReferences(templateFiles.freestyle.code)
+          : []),
+      ];
+
+      const seen = new Set();
+      const uniqueBlockReferences = allBlockReferences.filter((ref) => {
+        const key = `${ref.blockName}|${ref.variant}|${ref.version}`;
+        if (seen.has(key)) return false;
+        seen.add(key);
+        return true;
+      });
+
+      if (!uniqueBlockReferences.length) {
+        setIsBuildingReferences(false);
+        setBlockReferences([]);
+        return;
+      }
+
+      const blockZUIDs = new Set();
+      const directVariantZUIDs = new Set();
+      const versionedRefs = [];
+
+      uniqueBlockReferences.forEach((ref) => {
+        if (ref.variant) {
+          if (ref.version) {
+            const blockModel = blockModels?.find(
+              (block) => block.name === ref.blockName?.replace(".html", "")
+            );
+            if (blockModel) {
+              versionedRefs.push({
+                ref,
+                modelZUID: blockModel.ZUID,
+                itemZUID: ref.variant,
+              });
+            }
+          } else {
+            directVariantZUIDs.add(ref.variant);
+          }
+        } else if (ref.blockName) {
+          const blockModel = blockModels?.find(
+            (block) => block.name === ref.blockName.replace(".html", "")
+          );
+          if (blockModel) {
+            blockZUIDs.add(blockModel.ZUID);
+          }
+        }
+      });
+
+      const blockZUIDToBaseItem = new Map();
+      (
+        await Promise.allSettled(
+          Array.from(blockZUIDs).map(async (blockId) => {
+            const storeItems = Object.values(contentItems).filter(
+              (item) => item?.meta?.contentModelZUID === blockId
+            );
+            const items =
+              storeItems.length > 0
+                ? storeItems
+                : (await dispatch(fetchItems(blockId, { limit: 100 })))?.data ??
+                  [];
+            return [blockId, items];
+          })
+        )
+      )
+        .filter(
+          (result) =>
+            result.status === "fulfilled" && !!result.value?.[1]?.length
+        )
+        .forEach(({ value: [blockId, items] }) => {
+          const earliest = items.reduce((a, b) =>
+            a.meta.createdAt < b.meta.createdAt ? a : b
+          );
+          blockZUIDToBaseItem.set(blockId, earliest);
+        });
+
+      const blockNameToInfo = new Map();
+      blockModels?.forEach((model) => {
+        const baseItem = blockZUIDToBaseItem.get(model.ZUID);
+        if (baseItem) {
+          blockNameToInfo.set(model.name, { modelZUID: model.ZUID, baseItem });
+        }
+      });
+
+      const variants = [];
+      uniqueBlockReferences.forEach((ref) => {
+        if (ref.variant || !ref.blockName) return;
+
+        const info = blockNameToInfo.get(ref.blockName.replace(".html", ""));
+        if (!info) return;
+
+        if (ref.version) {
+          versionedRefs.push({
+            ref,
+            modelZUID: info.modelZUID,
+            itemZUID: info.baseItem.meta.ZUID,
+          });
+        } else {
+          variants.push(info.baseItem);
+        }
+      });
+
+      if (versionedRefs.length) {
+        (
+          await Promise.allSettled(
+            versionedRefs.map(async ({ ref, modelZUID, itemZUID }) => {
+              const storeVersions = contentVersions?.[itemZUID];
+              const storeVersion = Array.isArray(storeVersions)
+                ? storeVersions.find(
+                    (v) => v.meta.version === Number(ref.version)
+                  )
+                : null;
+
+              if (storeVersion) return storeVersion;
+
+              await dispatch(fetchVersions(modelZUID, itemZUID));
+
+              const freshVersions =
+                store.getState().contentVersions?.[itemZUID];
+              return (
+                freshVersions?.find(
+                  (v) => v.meta.version === Number(ref.version)
+                ) ?? null
+              );
+            })
+          )
+        )
+          .filter((result) => result.status === "fulfilled" && !!result.value)
+          .forEach(({ value }) => variants.push(value));
+      }
+
+      const directVariants = (
+        await Promise.allSettled(
+          Array.from(directVariantZUIDs).map(async (variantZUID) => {
+            if (!contentItems[variantZUID]) {
+              await dispatch(searchItems(variantZUID));
+            }
+            return store.getState().content[variantZUID] ?? null;
+          })
+        )
+      )
+        .filter((result) => result.status === "fulfilled" && !!result.value)
+        .map((result) => result.value);
+      variants.push(...directVariants);
+
+      setBlockReferences(variants);
+      setIsBuildingReferences(false);
+    };
+
+    run();
+  }, [templateFiles, blockModels]);
+
+  if (
+    !hasFields &&
+    !isFetchingFields &&
+    !isBuildingReferences &&
+    !blockReferences.length
+  ) {
     return (
       <Container disableGutters maxWidth="lg" sx={{ height: "100%" }}>
         <Stack p={4} height="100%" justifyContent="center">
@@ -141,6 +349,12 @@ export default function Content(props) {
               hasErrors={props.hasErrors}
               isLoadingItem={isLoadingItem}
             />
+            {props.model.type !== "block" && (
+              <UsedBlocks
+                blockReferences={blockReferences}
+                isBuildingReferences={isBuildingReferences}
+              />
+            )}
           </Box>
         </Box>
       </Box>
